+++ packages/core/forms/src/components/FormGenerator.vue
@@ -138,8 +138,8 @@
 
 <script>
 /**
- * @typedef {import('./types').FGCollapsibleOptions} FGCollapsibleOptions
- * @typedef {import('./types').FGSlots} FGSlots
+ * @typedef {import('../types').FGCollapsibleOptions} FGCollapsibleOptions
+ * @typedef {import('../types').FGSlots} FGSlots
  *
  * @typedef PartialGroup
  * @prop {FGCollapsibleOptions=} collapsible
