+++ packages/core/forms/src/components/fields/FieldArrayItem.vue
