+++ packages/core/forms/src/components/fields/FieldAutoSuggest.vue
@@ -44,7 +44,7 @@
 import { FORMS_API_KEY } from '../../const'
 import abstractField from './abstractField'
 import debounce from 'lodash-es/debounce'
-import { isValidUuid } from '../utils/isValidUuid'
+import { isValidUuid } from '../../utils/isValidUuid'
 
 const requestResultsLimit = 50
 const fieldStates = {
