+++ packages/core/forms/src/components/fields/FieldChecklist.vue
@@ -23,7 +23,7 @@ import clone from 'lodash-es/clone'
 import isObject from 'lodash-es/isObject'
 import isNil from 'lodash-es/isNil'
 import abstractField from './abstractField'
-import { slugify } from '../utils/schema'
+import { slugify } from '../../utils/schema'
 
 export default {
   mixins: [abstractField],
