+++ packages/core/forms/src/components/fields/abstractField.ts
@@ -7,8 +7,8 @@ import isFunction from 'lodash-es/isFunction'
 import isString from 'lodash-es/isString'
 import arrayUniq from 'lodash-es/uniq'
 import uniqueId from 'lodash-es/uniqueId'
-import { slugifyFormID } from '../utils/schema'
-import validators from '../utils/validators'
+import { slugifyFormID } from '../../utils/schema'
+import validators from '../../utils/validators'
 
 function convertValidator(validator) {
   if (isString(validator)) {
