+++ packages/core/forms/src/components/forms/schemas/OIDCAuth.js
