+++ packages/core/forms/src/components/forms/schemas/OIDCCommon.js
