--- packages/core/forms/src/generator/types.ts
@@ -1,16 +1,0 @@
-export type FGCollapsibleOptions = boolean | {
-  title?: string
-  description?: string
-  nestedCollapsible?: {
-    fields: any[],
-    triggerLabel: {
-      expand: string
-      collapse: string
-    }
-  }
-}
-
-export interface FGSlots {
-  beforeContent?: string
-  emptyState?: string
-}
