--- packages/core/forms/src/generator/utils/fieldsLoader.ts
@@ -1,20 +1,0 @@
-export { default as FieldAdvanced } from '../fields/FieldAdvanced.vue'
-export { default as FieldArray } from '../fields/FieldArray.vue'
-export { default as FieldArrayCardContainer } from '../fields/FieldArrayCardContainer.vue'
-export { default as FieldArrayItem } from '../fields/FieldArrayItem.vue'
-export { default as FieldArrayMultiItem } from '../fields/FieldArrayMultiItem.vue'
-export { default as FieldAutoSuggest } from '../fields/FieldAutoSuggest.vue'
-export { default as FieldCheckbox } from '../fields/FieldCheckbox.vue'
-export { default as FieldChecklist } from '../fields/FieldChecklist.vue'
-export { default as FieldInput } from '../fields/FieldInput.vue'
-export { default as FieldKeyValuePairs } from '../fields/FieldKeyValuePairs.vue'
-export { default as FieldLabel } from '../fields/FieldLabel.vue'
-export { default as FieldMetric } from '../fields/FieldMetric.vue'
-export { default as FieldMultiselect } from '../fields/FieldMultiselect.vue'
-export { default as FieldObject } from '../fields/FieldObject.vue'
-export { default as FieldObjectAdvanced } from '../fields/FieldObjectAdvanced.vue'
-export { default as FieldRadio } from '../fields/FieldRadio.vue'
-export { default as FieldSelect } from '../fields/FieldSelect.vue'
-export { default as FieldSelectionGroup } from '../fields/FieldSelectionGroup.vue'
-export { default as FieldSwitch } from '../fields/FieldSwitch.vue'
-export { default as FieldTextArea } from '../fields/FieldTextArea.vue'
