+++ packages/core/forms/src/types.ts
@@ -1,1 +1,16 @@
-export * from './generator/types'
+export type FGCollapsibleOptions = boolean | {
+  title?: string
+  description?: string
+  nestedCollapsible?: {
+    fields: any[],
+    triggerLabel: {
+      expand: string
+      collapse: string
+    }
+  }
+}
+
+export interface FGSlots {
+  beforeContent?: string
+  emptyState?: string
+}
