+++ packages/core/forms/src/utils/fieldsLoader.ts
@@ -1,0 +1,20 @@
+export { default as FieldAdvanced } from '../components/fields/FieldAdvanced.vue'
+export { default as FieldArray } from '../components/fields/FieldArray.vue'
+export { default as FieldArrayCardContainer } from '../components/fields/FieldArrayCardContainer.vue'
+export { default as FieldArrayItem } from '../components/fields/FieldArrayItem.vue'
+export { default as FieldArrayMultiItem } from '../components/fields/FieldArrayMultiItem.vue'
+export { default as FieldAutoSuggest } from '../components/fields/FieldAutoSuggest.vue'
+export { default as FieldCheckbox } from '../components/fields/FieldCheckbox.vue'
+export { default as FieldChecklist } from '../components/fields/FieldChecklist.vue'
+export { default as FieldInput } from '../components/fields/FieldInput.vue'
+export { default as FieldKeyValuePairs } from '../components/fields/FieldKeyValuePairs.vue'
+export { default as FieldLabel } from '../components/fields/FieldLabel.vue'
+export { default as FieldMetric } from '../components/fields/FieldMetric.vue'
+export { default as FieldMultiselect } from '../components/fields/FieldMultiselect.vue'
+export { default as FieldObject } from '../components/fields/FieldObject.vue'
+export { default as FieldObjectAdvanced } from '../components/fields/FieldObjectAdvanced.vue'
+export { default as FieldRadio } from '../components/fields/FieldRadio.vue'
+export { default as FieldSelect } from '../components/fields/FieldSelect.vue'
+export { default as FieldSelectionGroup } from '../components/fields/FieldSelectionGroup.vue'
+export { default as FieldSwitch } from '../components/fields/FieldSwitch.vue'
+export { default as FieldTextArea } from '../components/fields/FieldTextArea.vue'
